Update spark/ray example (#23)
This pull request updates the dependencies and configurations for the
Flyte plugins used in the `ray` and `spark` examples. The main changes
include adding new plugin dependencies, modifying environment variables,
and updating paths for executors and applications.

### Updates to dependencies:

- **`examples/plugins/ray_example.py`**: Added `flyteplugins-ray` to the
pip packages for the Ray plugin.
- **`examples/plugins/spark_example.py`**: Replaced `pyspark` with
`flyteplugins-spark` in the pip packages for the Spark plugin.
- **`plugins/ray/pyproject.toml`**: Added `flyte` as a dependency for
the Ray plugin.
- **`plugins/spark/pyproject.toml`**: Added `flyte` as a dependency for
the Spark plugin.

### Configuration updates:

- **`examples/plugins/spark_example.py`**: Updated paths for the
executor and application from `/opt/union/venv/bin` to `/opt/venv/bin`.

Signed-off-by: Kevin Su <pingsutw@apache.org>