Correct tag assignment in dev mode (#63)
`make dist` invokes setuptools_scm's version scheme, which adds a hash
based on the commits between the latest commit, and the latest tagged
version. However, the version it produces doesn't take into account
files that may have changed (ie git is dirty). We used to just hash the
whl file itself - this behavior was circumvented when adding the logic
for the preset tag but we'll revert to it now by not setting the tag
when we're in dev mode.

We need the image tag to change because even if docker cache doesn't
kick in, these images are often run with IfNotPresent pull policy -
which means even though you've pushed an updated version locally as part
of your dev cycle, the Pod that kicks off might be using the older image
and you'd be none the wiser.

---------

Signed-off-by: Kevin Su <pingsutw@apache.org>
Signed-off-by: squiishyy <joe@union.ai>