Wheel changes (#65)
Updated version of https://github.com/flyteorg/flyte-sdk/pull/36

This changes how python wheels are handled... currently we're just
installing *whl, but this doesn't work for multi-arch wheels. pip knows
how to find the correct architecture with the --find-links arg, but we
need to tweak the command. However, we want
* if there's a local wheel, we absolutely want to use it, even if
there's a newer version on pypi. So the first command doesn't install
any dependencies and uses --no-index.
* Since we need the dependencies, we need a second command to install
the rest of the deps from pypi.

Will need to update the remote builder as well.

> [!NOTE]
> Had to update the image identifier in the unit test because the layer
definition string representation actually changed.

---------

Signed-off-by: Yee Hing Tong <wild-endeavor@users.noreply.github.com>
Signed-off-by: squiishyy <joe@union.ai>