Remove editable (#76)
This removes the pretty hacky editable install detection and now relies
purely on detecting the 'dev' word in the version that should be placed
there by setuptools_scm.

The recent change to again hash the built wheel when constructing local
images broke image building for releases because the missing preset tag
caused the image to be tagged with a hash instead of the release tag.
example:
https://github.com/flyteorg/flyte-sdk/actions/runs/16977992633/job/48132152647

b9 images have been built and pushed locally.

Signed-off-by: Yee Hing Tong <wild-endeavor@users.noreply.github.com>
Signed-off-by: squiishyy <joe@union.ai>