Example: parameterizing image uris for flyte run/deploy via env vars (#88)
The example in this PR shows how to parameterize image uris via
environment variables.

The use case for this is as follows:
- The user creates an image from an out-of-band process, either locally
or through CI/CD. The assumption is that they have some pre-built uri
that they want to use with the task environment.
- The user wants to make zero code changes but deploy the task with a
different image for dev/staging/production

In this example, the user needs to provide the `BASE_IMAGE` environment
wherever they run the `flyte run/deploy` command. By passing this
environment variable to the `env` argument, `TaskEnvironment(...,
env={...})`, `BASE_IMAGE` will be propagated to all the invocations of
tasks in that TaskEnvironment.

---------

Signed-off-by: Niels Bantilan <niels.bantilan@gmail.com>