[Fix] Image Cache bug when python version different between local and remote (#163)
## Description

When the image has the version that's different from local env (e.g. use
python 3.10 image but use 3.12 in local env), the remote execution will
fail as it cannot find the image.

## Changes

Exclude python version when creating image identifier hash. Change the
structure of look up table into nested. Below is the example:

```sh
{'rRtF_Uo8Wabpz3pMlhjELA': {'3.10': '356633062068.dkr.ecr.us-east-2.amazonaws.com/union/playground:flyte-7a4ec1d4c17a82d32086f976606d13ab'}}
```

If the running env use different environment (e.g. 3.11), which is not
in the look-up table, it will fall back to the first image, i.e. "3.10",
in the dict

---------

Signed-off-by: machichima <nary12321@gmail.com>