Introduce parent_env_name which is pickable (#236)
The move to using tasks' environment name as the lookup key in the image
cache, introduced by #211, doesn't work for pickled tasks, because that
field is
[removed](https://github.com/flyteorg/flyte-sdk/blob/bd975bf43e70b39f7a6003104f99ce90c326d69f/src/flyte/_task.py#L150)
during the pickle process. This leads to the "Task template has no
parent environment" error.

This adds a string field that should always be present for correct
lookup.

---------

Signed-off-by: Yee Hing Tong <wild-endeavor@users.noreply.github.com>