refactor(models): Use get_type_hints for param and return types (#295)
We currently use `inspect.signature` when to read types at run time. But
this breaks if delayed annotations is enabled (from future import
annotations). This adds `typing.get_type_hints` to help when we call
`from_callable`.

However, `typing.get_type_hints` does not work for types that aren't
defined at the module level. It doesn't capture closures and attempting
to do so is something we can pursue in the future.

---------

Signed-off-by: Kevin Su <pingsutw@apache.org>
Signed-off-by: Yee Hing Tong <wild-endeavor@users.noreply.github.com>
Co-authored-by: Yee Hing Tong <wild-endeavor@users.noreply.github.com>