[Feature] Add task git repo host uri when deploying task (#335)
## Changes
We will store flyte task git repo host url in the backend when sending a
deploy request request to the run service. The url will be respond in
the GetTaskDetails api.

How the url is defined:
1. If task code file is under git repo and the git tree is clean, the
url will be suffixed with line number.
2. If task code file is under git repo and the git tree is not clean,
the url will only points to the code file in remote repo.
3. If task code file is not under git repo, the url will not be sent to
the backend.
4. The SDK will choose origin as default remote. If origin remote does
not exist, the SDK will sort remotes alphabetically and choose the first
one.

Signed-off-by: Alex Wu <c.alexwu@gmail.com>
Signed-off-by: “Alex <alexchien130@gmail.com>